Create Installer.bat
This is really the meat of the project. I'm putting it here for transparency but I will keep distributing as a .exe since that is more user friendly and allows me to pack all the assets inside. I'm not uploading the assets yet as I still intend to make a few more tweaks so this bat won't actually work yet. In time this will be usable and all dependencies will either be in this repo or linked in submodules for  fully modular installation